Adding project files
To make it easier to work with IDE (Geppetto in my case) I added a .project file to the project (which is a standard practice in Puppet) and added missing dependencies in metadata.json.

The dependencies to puppetlabs/apache and puppetlabs/java are actually optional (and only used in tests) but there is no way as far as I know to mark dependencies as optional in metadata.json. I'm not completely sure what is the best approach here.