Adding a `$ensure` parameter to agents_package
This is done as a proof of concept, same approach could / should be replicated to other resources.

The idea is that to keep the declarative nature of Puppet, each defined type should manage the full lifecycle of the resources it manages. This enable us to remove a lot of `if`s in the code and just pass the correct value to `$ensure` instead.

Let me know if you find this idea useful or not. Depending on your answer, I can replicate this mechanism to most resources in puppet-dynatrace.